implicitly cast pthread_t to long unsigned int
This fixes compilation time errors like:

	error: format '%lu' expects argument of type 'long unsigned int', but argument 5 has type 'pthread_t' {aka 'struct __pthread *'}

+++ libtcmu.c
@@ -860,7 +860,7 @@ void tcmu_set_thread_name(const char *prefix, struct tcmu_device *dev)
 
 	if (!prefix) {
 		tcmu_dev_err(dev, "Failed to set name for thread %lu\n",
-			     pthread_self());
+			     (long unsigned int)pthread_self());
 		return;
 	}
